+++ Genesis/src/utopia/genesis/handling/template/EventHandler2.scala
@@ -1,0 +1,31 @@
+package utopia.genesis.handling.template
+
+import utopia.inception.util.Filter
+
+/**
+  * Event handlers distribute events for the handleable instances
+  * @author Mikko Hilpinen
+  * @since 10.5.2019, v1+
+  */
+trait EventHandler2[Listener <: Handleable2, -Event] extends Handler2[Listener]
+{
+	/**
+	  * Finds an event filter for the specified listener
+	  * @param listener A listener
+	  * @return An event filter used by the specified listener
+	  */
+	protected def filterOf(listener: Listener): Filter[Event]
+	
+	/**
+	  * Informs a single listener about an event
+	  * @param listener A listener
+	  * @param event An event
+	  */
+	protected def deliver(listener: Listener, event: Event): Unit
+	
+	/**
+	  * Distributes an event between the listeners
+	  * @param event An event
+	  */
+	def distribute(event: Event): Any = items.foreach { l => if (filterOf(l)(event)) deliver(l, event) }
+}