+++ Genesis/src/utopia/genesis/handling/template/Handleable2.scala
@@ -1,0 +1,19 @@
+package utopia.genesis.handling.template
+
+import utopia.flow.view.template.eventful.FlagLike
+
+/**
+  * Common trait for items which may be placed within handlers
+  * @author Mikko Hilpinen
+  * @since 30/01/2024, v3.6
+  */
+trait Handleable2
+{
+	// ABSTRACT ---------------------
+	
+	/**
+	  * @return A condition that must be met in order for this item to be handled.
+	  *         The [[Handler2]] implementations *should* respect this condition.
+	  */
+	def handleCondition: FlagLike
+}