+++ Genesis/Changes.md
@@ -1,5 +1,8 @@
 # Utopia Genesis - List of Changes
 
+## v3.6 (in development)
+Handler update
+
 ## v3.5.1 (in development)
 ### New methods
 - **StrokeSettings**